+++ index.html
@@ -2,8 +2,6 @@
 <html>
   <head>
     <title>2DoBox</title>
-    <link rel="stylesheet" href="css/reset.css" type="text/css">
-    <link rel="stylesheet" href="css/styles.css" type="text/css">
     <link href="https://fonts.googleapis.com/css?family=Roboto+Slab:400,700" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css?family=Open+Sans" rel="stylesheet">
     <meta name="viewport" content="width=device-width, initial-scale=1, user-scalable = true">
@@ -29,6 +27,5 @@
 
   </body>
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.12.4/jquery.min.js"></script>
-    <script type="text/javascript" src="javascript.js"></script>
     <script src="main.bundle.js"></script>
 </html>
