+++ package.json
@@ -34,6 +34,7 @@
     "node-sass": "^3.13.0",
     "sass-loader": "^4.0.2",
     "style-loader": "^0.13.1",
+    "svg-url-loader": "^1.1.0",
     "webpack": "^1.13.3",
     "webpack-dev-server": "^1.16.2"
   }
